Add files via upload
This is a very rough first attempt at generating visualizations in the complex hexapolar plane, and in this case visualizations of variations on the Mandelbrot function. You can focus on one orthant at a time or stitch together four-quadrant projections. You can also toggle the modulus between natural (i.e., zz*) and thumb-on-the-scale Euclidean. Functions for computing consolidation (addition generalized such that it is not necessarily associative), multiplication, exponentiation (to integer powers), and conjugation of hexapolar or complex hexapolar numbers are included in the page source. Because the complex hexapolar plane embeds/extends the complex plane, you can ignore the hexapolar stuff (more or less) and focus on the complex plane if desired.