Create consolidate.js
Addition is associative and commutative. Addition generalized such that it is not necessarily associative may be referred to as "consolidation." This function accepts sign and magnitude arguments for two complex numbers and returns signs and magnitudes for their "sum" or "consolid." It can be used with the complex hexapolar number system or a multipolar number system that does not embed the real line or complex plane.