Create cHexNaturalExponent.js
Exponentiation as repeated multiplication inspired this function which accepts signs and magnitudes for components of a complex hexapolar number and a natural number as a power to which it is raised (i.e., how many times it is multiplied by itself). In its current state, that natural number must be >; signs may be expressed as even exponents of s modulo 12.