Create cHexCosSin.js
This function accepts an argument in radians and a cycle of exponentiation (one of three) and returns the signs and magnitudes of the components of a complex hexapolar number; it extrapolates the behavior of an imaginary-like unit under repeated self-multiplication in contrast to an exponential function defined via power series (see cHexPowerSeries.js). 

Some numbers in the complex hexapolar plane do not have square roots in the complex hexapolar plane; those that do have square roots may be outputs of this function depending on the cycle of exponentiation to which they belong (the complex hexapolar plane contains three such cycles canonically oriented to exponentiation of s, s^3, and s^5).