Create cHexExpPowerSeries.js
This function generalizes the complex exponential function to complex hexapolars as per a power series definition such that the non-associativity of consolidation is compensated for by pairwise consolidation of terms with each iteration. It accepts a complex hexapolar number and a maximum number of iterations. Nine iterations are sufficient to generate an approximation of e or e^-1 given a real input; after ~200 iterations, numerical precision is strained with diminishing returns.  

Real and complex inputs have expected outputs; nonpositive real-like hexapolar unit inputs converge to one of two quantities other than e^-1, depending on sign subset. 

Generalizing exp(i(theta)) to exp(s(theta)) returns results that differ from extrapolation of natural exponentiation or repeated multiplication of an imaginary-like hexapolar unit such that differences are most evident if theta is not an integer multiple of pi/2.