Add files via upload
Corrected graphs of a generalized exponential function defined with power series; x axis is argument expressed as a multiple of pi and y axis is modulus such that black data points correspond to an input of i, -i (s^3 or s^9), red data points to the Euclidean modulus of other powers of s, and purple data points to the natural (conjugation) modulus of other powers of s (for black data points both moduli are equal). The corrected function still produces nonnumeric results as the argument approaches 36pi and appears unstable prior as depicted in the graph.